feat: centralise Amplify client mock
